refactor(discord): #4229 S6 — extract no-result exit handler (move-pure)
Move the watcher loop's no-result exit block C (tmux_watcher.rs L557-1272,
716 lines) verbatim into tmux_watcher/no_result_exits.rs behind a
handle_no_result_exits(context, locals, state) -> NoResultExitOutcome seam:
fresh ready-for-input idle structural finalize, tmux-death break,
cancel/shutdown break, ready-for-input stall notice + dispatch-fail, and the
#3419 watchdog-timeout finalize.

Behavior-preserving decompose: the moved body is byte-identical apart from the
mechanical seam substitutions (12 continue -> ContinueWatcherLoop, 2
break 'watcher_loop -> BreakWatcherLoop, Fallthrough tail; state-field
&mut/deref rewrites; current_offset shadow + single write-back). Threaded
context=11, locals=13, state=16.

Parent ratchet 5007 -> 4351 prod LoC; new module registered under the
tmux_watcher/** namespace cap (814 LoC, below the 1000 giant threshold).
Regenerated inventory docs.
